photoURL is erasing
Every time I update my profile, the photoURL is erasing. So I found out that component "user" is calling this function with null input and here we have no check if this url was null or not.
So checking !photoURL and not updating in case of null will solve the problem.